+++ CHANGELOG.md
@@ -8,6 +8,7 @@ Subheadings to categorize changes are `added, changed, deprecated, removed, fixe
 
 ### added
 
+- added `AsyncTask::with_timeout(mut self, dur)`
 - added timeout support via `AsyncTask::new_with_timeout(dur, f)`
 - added `AsyncTask::pending()` as an abstraction for a task that never finishes